movie-website/pyproject.toml
[tool.poetry]
name = "movie-website"
version = "0.1.0"
description = ""
authors = ["Momayaz <momayaz.dr.1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
Pillow = "^8.1.2"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
